Extract PDF panel rendering modules
Split page-level PDF rendering into focused panel modules while keeping pdf_page1.py and pdf_page2.py as thin composers with stable entry points.
