+++ apps/server/tests/adapters/pdf/test_report_pdf_rendering.py
@@ -18,10 +18,10 @@ from test_support.report_helpers import (
 )
 
 from vibesensor import __version__
+from vibesensor.adapters.pdf._panel_systems import _draw_system_card
 from vibesensor.adapters.pdf.mapping import map_summary
 from vibesensor.adapters.pdf.pdf_diagram_render import car_location_diagram
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
-from vibesensor.adapters.pdf.pdf_page1 import _draw_system_card
 from vibesensor.adapters.pdf.report_data import PartSuggestion, SystemFindingCard
 from vibesensor.shared.constants import KMH_TO_MPS
 from vibesensor.use_cases.diagnostics import summarize_log
