+++ apps/server/vibesensor/adapters/pdf/__init__.py
@@ -14,7 +14,7 @@ Module topology
 - **Mapping**: ``mapping.py`` (thin mapper: context → ``ReportTemplateData``).
 - **Mapping helpers**: ``peak_table.py`` (peak rows), ``report_sections.py`` (report sections), ``presentation.py`` (render-only labels), ``pattern_parts.py`` (parts suggestions).
 - **Engine**: ``pdf_engine.py`` (public entry, page orchestration, pagination).
-- **Pages**: ``pdf_page1.py``, ``pdf_page2.py``.
+- **Pages**: ``pdf_page1.py``, ``pdf_page2.py`` (thin page composers over ``_panel_*.py`` modules).
 - **Diagrams**: ``pdf_diagram_render.py`` (layout planning, drawing).
 - **Primitives**: ``pdf_drawing.py``, ``pdf_text.py``,
   ``pdf_style.py`` (theme, constants, page geometry, render context).
